requirements.txt
altgraph==0.17.4
future==1.0.0
iso8601==2.1.0
macholib==1.16.3
macos-notifications==0.2.1
packaging==25.0
psutil==7.0.0
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyserial==3.5
pync==2.0.3
PyYAML==6.0.2
